# uint256: optimize Mod, DivMod, AddMod